Update installation docs to focus on system installs
The suggested command to extract the tarball now places it under `/opt`
without having to manually move it, and use of `sudo` ensures that the
root account owns the files. The second half is particularly important
since the script will normally get run as root, so we want to be sure
that random users aren't able to edit the script maliciously!